Copy Tile method moved to specific auxiliar class
